feat(F-58): Adjustments
