+++ frontend/src/app/[locale]/(main)/training-settings/page.tsx
@@ -1,14 +1,7 @@
-import type { Metadata } from 'next';
-import TrainingSettingsClient from './TrainingSettingsClient';
-
-export const metadata: Metadata = {
-  title: 'CoachAI - Training Settings',
-  description: 'Customize your training preferences and goals',
-};
-
 import { generateMetadata as generateDynamicMetadata } from '@/lib/metadata';
 import type { Metadata } from 'next';
 import type { Props } from '@/interfaces/LayoutProps';
+import TrainingSettings from '../../../../components/common/TrainingSettings';
 
 export async function generateMetadata({ params }: Props): Promise<Metadata> {
   const { locale } = await params;
@@ -16,9 +9,5 @@ export async function generateMetadata({ params }: Props): Promise<Metadata> {
 }
 
 export default function TrainingSettingsPage() {
-  return (
-    <div className="max-w-4xl mx-auto container">
-      <TrainingSettingsClient />
-    </div>
-  );
+  return <TrainingSettings />;
 }
